now test uses ClockData

+++ src/fitbit/DataExpert.java
@@ -22,7 +22,7 @@ public class DataExpert {
 //        UserData.setSex(s);
     }
 
-    int getTimeData(){
+    double getTimeData(){
         return ClockData.getTimeData();
     }
 }
